requirements.txt
asgiref==3.9.0
certifi==2025.7.9
charset-normalizer==3.4.2
cloudinary==1.44.1
Django==5.2.4
django-cloudinary-storage==0.3.0
django-environ==0.12.0
django-filter==25.1
django-ratelimit==4.1.0
djangorestframework==3.16.0
idna==3.10
pillow==11.3.0
requests==2.32.4
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.14.1
tzdata==2025.2
ulid-py==1.1.0
urllib3==2.5.0
